+++ crates/errors/src/errors.rs
@@ -1,6 +1,7 @@
 use drv_errors_derive::DrvError;
 use drv_models::state::types::{account_type::AccountType, AssetType, OrderSide, TokenProgram};
 use serde::{Deserialize, Serialize};
+use solana_program::pubkey::PubkeyError;
 #[cfg(feature = "on-chain")]
 use solana_program::{msg, program_error::ProgramError, pubkey::Pubkey};
 #[cfg(feature = "off-chain")]
@@ -664,6 +665,18 @@ pub enum DeriverseErrorKind {
         msg = "Provided variance is too small, variance {variance} < min {min_variance}"
     )]
     InvalidVariance { variance: f64, min_variance: f64 },
+
+    #[error(code = 293, msg = "Airdrop amount must be > 0")]
+    InivalidAirdrop { wallet_address: Pubkey, ratio: f64 },
+
+    #[error(code = 294, msg = "Airdrop authority wasnt initlised correctly")]
+    InvalidAirdropAuthority { root_address: Pubkey },
+
+    #[error(
+        code = 295,
+        msg = "Invalid Spl Token Program Id, expected: {expected}, actual: {actual}"
+    )]
+    InvalidSplTokenProgramId { expected: Pubkey, actual: Pubkey },
 }
 
 #[cfg(test)]