+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 
+## [v2.8] - 2025-11-12
+
+## Added
+- `airdrop_authority_address` to `RootState`
+- `ChangeAirdropAuthority` instruction
+
 ## [v2.7] - 2025-11-12
 
 ## Changed